Move object decoding into decodeobjects module
