+++ packages/iov-bns/src/bnsconnection.ts
@@ -52,18 +52,17 @@ import { Stream, Subscription } from "xstream";
 import { bnsCodec } from "./bnscodec";
 import { swapToAddress } from "./conditions";
 import { ChainData, Context } from "./context";
+import { decodePubkey, decodeUserData } from "./decode";
 import {
+  decodeAmount,
   decodeCashConfiguration,
   decodeElectionRule,
   decodeElectorate,
   decodeProposal,
-  decodePubkey,
   decodeToken,
-  decodeUserData,
   decodeUsernameNft,
-} from "./decode";
-import { decodeVote } from "./decodemsg";
-import { decodeAmount } from "./decodinghelpers";
+  decodeVote,
+} from "./decodeobjects";
 import * as codecImpl from "./generated/codecimpl";
 import { bnsSwapQueryTag } from "./tags";
 import {
