+++ packages/iov-bns/types/decodemsg.d.ts
@@ -1,6 +1,3 @@
 import { UnsignedTransaction } from "@iov/bcp";
-import * as codecImpl from "./generated/codecimpl";
-import { BnsdTxMsg, Keyed, Vote } from "./types";
-import { IovBech32Prefix } from "./util";
-export declare function decodeVote(prefix: IovBech32Prefix, vote: codecImpl.gov.IVote & Keyed): Vote;
+import { BnsdTxMsg } from "./types";
 export declare function decodeMsg(base: UnsignedTransaction, tx: BnsdTxMsg): UnsignedTransaction;
